bioisac-poc/requirements.txt
slack-bolt==1.22.0
slack-sdk==3.33.5
mysql-connector-python==9.1.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
python-dateutil==2.9.0.post0
